@@ -8,7 +8,7 @@
 	</PropertyGroup>
 
 	<PropertyGroup>
-		<TargetFrameworks>netstandard2.0;netcoreapp3.1;net6.0;net7.0;net8.0;net9.0</TargetFrameworks>
+		<TargetFrameworks>netstandard2.0;net6.0;net7.0;net8.0;net9.0;net10.0</TargetFrameworks>
 		<SignAssembly>true</SignAssembly>
 		<AssemblyOriginatorKeyFile>..\.Documents\Forge.pfx</AssemblyOriginatorKeyFile>
 		<ProjectGuid>{1D1A1570-3828-4C96-A040-6E705C5EF31F}</ProjectGuid>
@@ -18,7 +18,7 @@
 	<PropertyGroup>
 		<Product>Forge: reference, practice and patterns implementations and helper(s)</Product>
 		<Company>Forge Project Contributor(s)</Company>
-		<Copyright>Copyright © Zoltan Juhasz, 2022-2025</Copyright>
+		<Copyright>Copyright © Zoltan Juhasz, 2022-2026</Copyright>
 		<PackageId>Forge.Security.Jwt.Client.Storage.Browser</PackageId>
 		<Title>Forge Security Jwt Client Storage Browser</Title>
 		<Description>Forge Security Jwt Web Browser storages</Description>
@@ -27,7 +27,7 @@
 		<Authors>Zoltan Juhasz</Authors>
 		<PackageReadmeFile>README.md</PackageReadmeFile>
 		<RepositoryType>git</RepositoryType>
-		<AssemblyVersion>1.3.6.0</AssemblyVersion>
+		<AssemblyVersion>1.4.0.0</AssemblyVersion>
 		<Version>$(AssemblyVersion)</Version>
 		<FileVersion>$(AssemblyVersion)</FileVersion>
 		<PackageLicenseFile>LICENSE</PackageLicenseFile>
@@ -39,7 +39,7 @@
 		<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
 	</PropertyGroup>
 
-	<ItemGroup Condition="'$(TargetFramework)' == 'net6.0' or '$(TargetFramework)' == 'net7.0' or '$(TargetFramework)' == 'net8.0' or '$(TargetFramework)' == 'net9.0'">
+	<ItemGroup Condition="'$(TargetFramework)' == 'net6.0' or '$(TargetFramework)' == 'net7.0' or '$(TargetFramework)' == 'net8.0' or '$(TargetFramework)' == 'net9.0'  or '$(TargetFramework)' == 'net10.0'">
 		<SupportedPlatform Include="browser" />
 	</ItemGroup>
 
